project structure is done + Custom404 is done
