+++ src/pages/404.tsx
@@ -1,0 +1,1 @@
+export { default } from './Custom404'
