+++ src/pages/Custom404/Custom404.tsx
@@ -1,0 +1,3 @@
+export default function Custom404() {
+  return <h1>404 - Page Not Found</h1>
+}
