+++ src/pages/Custom404/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Custom404'
